gmail-to-s3/requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
boto3>=1.28.0
aws-durable-execution-sdk-python
